+++ examples/apiserver/BUILD
@@ -16,7 +16,6 @@ go_library(
         "//cmd/libs/go2idl/client-gen/test_apis/testgroup/v1:go_default_library",
         "//examples/apiserver/rest:go_default_library",
         "//pkg/api:go_default_library",
-        "//pkg/genericapiserver/authorizer:go_default_library",
         "//pkg/genericapiserver/registry/generic:go_default_library",
         "//pkg/genericapiserver/registry/rest:go_default_library",
         "//pkg/genericapiserver/server:go_default_library",
@@ -26,6 +25,7 @@ go_library(
         "//vendor:github.com/golang/glog",
         "//vendor:k8s.io/apimachinery/pkg/runtime/schema",
         "//vendor:k8s.io/apimachinery/pkg/util/errors",
+        "//vendor:k8s.io/apiserver/pkg/authorization/authorizerfactory",
     ],
 )
