+++ staging/src/k8s.io/apiserver/pkg/authentication/authenticatorfactory/delegating.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package authenticator
+package authenticatorfactory
 
 import (
 	"errors"
