+++ staging/src/k8s.io/apiserver/pkg/authorization/authorizerfactory/authz_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package authorizer
+package authorizerfactory
 
 import (
 	"testing"
